google_cloud_functions/cloud_vision_test/requirements.txt
google-cloud-storage
google-cloud-vision
google-cloud-datastore
google-cloud-firestore
